pyproject.toml
[tool.poetry]
name = "nextcloudappstore"
version = "4.0.0"
description = "App store for Nextcloud apps"
authors = ["Morris Jobke <hey@morrisjobke.de>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "~3.2"
django-braces = "~1.15.0"
djangorestframework-camel-case = "^1.3.0"
django-parler-rest = "^2.1"
django-parler = "^2.3"
django-allauth = "^0.54.0"
lxml = "^4.7.1"
requests = ">=2.27.0, <2.30.0"  # till cachecontrol get updated
django-cors-middleware = "^1.5.0"
django-csp = "^3.7"
pyOpenSSL = "^23.0.0"
Markdown = "^3.3.6"
bleach = "^6.0.0"
pymple = "^0.1.3"
semantic-version = "^2.8.5"
pem = "^21.2.0"
django-simple-captcha = "^0.5.14"
uWSGI = "^2.0.20"
newrelic = "^7.2.4"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-cov = "*"
pytest-django = "*"
responses = "*"
django-debug-toolbar = "*"
coverage = "*"
selenium = "*"
coveralls = "*"
types-requests = "^2.26.3"
types-pyOpenSSL = "^23.0.0"

[build-system]
requires = ["poetry>=1.3.1"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 120
target-versions = ["py38"]
preview = true

[tool.isort]
profile = "black"
